Working on path following.  Not yet tested, but starting to come together

+++ trunk/src/common/mograph.hpp
@@ -67,7 +67,7 @@ namespace Skeletal
     Motion random_motion(int l) const;
     
     //Synthesizes a motion which follows a path
-    Motion follow_path(Vector2d (*path_func)(double), double max_d, double dur) const;
+    Motion follow_path(Vector2f (*path_func)(double), double max_d, double dur) const;
     
     //Extracts a linear submotion from the motion graph
     Motion submotion(int start, int end) const;
